GIX-995: Sns Wallet Details Setup
