+++ frontend/src/lib/constants/routes.constants.ts
@@ -6,7 +6,8 @@ export enum AppPath {
   Neurons = "/#/u/:rootCanisterId/neurons",
   Proposals = "/#/proposals",
   Canisters = "/#/canisters",
-  Wallet = "/#/wallet",
+  LegacyWallet = "/#/wallet",
+  Wallet = "/#/u/:rootCanisterId/wallet",
   ProposalDetail = "/#/proposal",
   LegacyNeuronDetail = "/#/neuron",
   CanisterDetail = "/#/canister",
