+++ frontend/src/lib/pages/SnsWallet.svelte
@@ -1,0 +1,2 @@
+<!-- TODO: GIX-995 Implement SNS Wallet details -->
+<div data-tid="sns-wallet">TO BE IMPLEMENTED</div>
